requirements.txt
absl-py==2.1.0
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
ase==3.22.1
astunparse==1.6.3
binaryornot==0.4.4
blinker==1.7.0
boto3==1.34.73
botocore==1.34.73
branca==0.7.1
bravado==11.0.3
bravado-core==6.1.1
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
chatbot==1.5.2b0
click==8.1.7
comtypes==1.2.0
cookiecutter==2.3.0
cryptography==42.0.3
datasets==2.15.0
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
docopt==0.6.2
evaluate==0.4.1
fastapi==0.109.2
ffmpy==0.3.1
flatbuffers==24.3.25
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
getorg==0.3.1
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.18.0
gradio_client==0.10.0
graphviz==0.20.1
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.3
hyperframe==6.0.1
imageio==2.34.0
ipyleaflet==0.18.2
ipython-genutils==0.2.0
jax==0.4.25
jmespath==1.0.1
jsonpointer==2.0
jsonref==1.1.0
kaggle==1.5.16
keras==2.12.0
keyboard==0.13.5
latexcodec==2.0.1
lazy_loader==0.3
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
matminer==0.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
monty==2023.11.3
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
munkres==1.1.4
neptune==1.3.1
nltk==3.8.1
nominatim==0.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai==1.12.0
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
orjson==3.9.13
p-tqdm==1.4.0
palettable==3.3.3
pandas==1.5.3
pathos==0.3.1
peewee==3.17.1
pipreqs==0.4.13
plotly==5.18.0
ply==3.11
portalocker==2.8.2
pox==0.3.3
ppft==1.7.6.7
protobuf==4.25.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybtex==0.24.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pyDOE==0.3.8
pydub==0.25.1
PyGithub==2.2.0
PyJWT==2.8.0
pymatgen==2023.11.12
pymongo==4.6.0
PyNaCl==1.5.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
python-multipart==0.0.9
python-pptx==0.6.23
python-slugify==8.0.1
pywinauto==0.6.6
qdrant-client==1.7.3
qudida==0.0.4
rapidfuzz==3.2.0
regex==2023.10.3
requests-oauthlib==2.0.0
responses==0.18.0
retrying==1.3.4
rich==13.5.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
ruff==0.2.1
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.0
scikit-image==0.22.0
semantic-version==2.10.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.19.2
SMACT==2.5.3
smmap==5.0.1
spglib==2.1.0
starlette==0.36.3
streamlit==1.32.1
swagger-spec-validator==3.0.3
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
text-unidecode==1.3
tifffile==2024.2.12
timm==0.9.6
tk==0.1.0
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1
torch_geometric==2.4.0
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
torchviz==0.0.2
traittypes==0.2.1
transformers==4.32.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
uncertainties==3.1.7
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
XlsxWriter==3.2.0
xxhash==3.4.1
xyzservices==2023.10.1
yarg==0.1.9
yarl==1.9.3
yfinance==0.2.36
